feat(ui): move UI components to 'packages/ui' directory
- Move 'separator.tsx', 'popover.tsx', 'button.tsx' and 'scroll-area.tsx' from 'apps/next/components/ui' to 'packages/ui/src'
- Update import paths in the Next.js app to reflect the new directory structure
